requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
pillow
numpy
pandas
scikit-learn
albumentations
tensorflow
python-dotenv
mlflow
prometheus-client
